RavenDB-9950 - Fixing the way we do binding to Sodium so we'll use API generation (adding a new API shouldn't take an hour). - Moving all the API to the new model - Entropy usage in SecretProtection is exactly 24 bytes now

+++ src/Raven.Server/Documents/AttachmentsStorageHelper.cs
@@ -15,7 +15,7 @@ namespace Raven.Server.Documents
             using (context.GetManagedBuffer(out JsonOperationContext.ManagedPinnedBuffer buffer))
             using (context.GetManagedBuffer(out JsonOperationContext.ManagedPinnedBuffer cryptoState))
             {
-                if(cryptoState.Length < Sodium.crypto_generichash_statebytes())
+                if(cryptoState.Length < (int)Sodium.crypto_generichash_statebytes())
                     throw new InvalidOperationException("BUG: shouldn't happen, the size of a generic hash state was too large!");
 
                 InitComputeHash(cryptoState);
